requirements-test.txt
coveralls
django-redis
openwisp-utils[qa]>=0.5.0
redis
channels-redis

